fastapi_blockchain_service/requirements.txt
fastapi
uvicorn
web3
pydantic
python-dotenv
requests 